Combine all chipdbs into yowasp-nextpnr-{ice40,ecp5} packages.
The PyPI package size limit used to be 60 MB, which is why the iCE40
target had to use split chipdb packages. (The iCE40 chipdbs are about
70 MB in size). The ECP5 target inherited this architecture.

The PyPI package size limit is now 100 MB, so everything can be
combined into a single package, eliminating most common installation
and dependency related issues at the cost of a small increase in
download size.